feat(Instagram): Integrate `Disable onboarding permission prompts` patch into `Disable analytics`
Also removed the old experiment-flag and legacy consent-URI approaches
